requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Config
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg[binary]==3.2.1

alembic==1.12.1

# PDF processing
PyMuPDF==1.23.8

# Embedding model (Amharic)
sentence-transformers==2.2.2

# Vector search
faiss-cpu==1.13.2

# Generator model (Walia-LLM)
transformers==4.36.0
torch==2.1.0
accelerate==0.25.0

# Utilities
numpy==1.26.4
pandas==2.0.3
loguru==0.7.2
tqdm==4.66.1
